feat(ui): extract remaining pure utils  (#28791)
* feat(ui): extract remaining pure utils (PR #12)

Extract pure functions from 12 utility files into dedicated *PureUtils.ts files.
Original files re-export all moved symbols for backward compatibility.

New pure util files:
- AgentsStatusWidgetPureUtils.ts (getAgentLabelFromType, getAgentStatusLabelFromStatus, getAgentStatusSummary)
- Alerts/AlertsUtilPure.ts (20 pure alert helpers)
- BlockEditorPureUtils.ts (pure block editor utils)
- ContainerDetailPureUtils.ts (pure container detail utils)
- ContextCenterPureUtils.ts (pure context center utils)
- CuratedAssetsPureUtils.ts (pure curated assets utils)
- CustomizableLandingPagePureUtils.ts (pure landing page utils)
- CustomizePage/CustomizePagePureUtils.ts (pure customize page utils)
- DataAssetSummaryPanelPureUtils.ts (pure data asset summary utils)
- CSV/CSVPureUtils.ts (pure CSV processing utils)
- Lineage/LineagePureUtils.ts (pure lineage utils)
- MetricEntityUtils/MetricPureUtils.ts (pure metric utils)
- UsersPureUtils.ts (pure user utils)


* fix checkstyle

* fix: drop dead duplicate code from BlockEditorPureUtils

Remove isHTMLString, _convertMarkdownFormatToHtmlString, formatContentForClient
from BlockEditorPureUtils.ts — all three are duplicates of code in BlockEditorUtils.ts,
none were re-exported or consumed anywhere. Pure file now contains only
isDescriptionContentEmpty and getTextFromHtmlString.


* fix(test): update DataAssetSummaryPanelUtils test to mock TablePureUtils

getTierTags moved to TablePureUtils.ts; test was mocking TableUtils (old path)
causing the mock to be ignored and the assertion to fail.


---------
