+++ openmetadata-ui/src/main/resources/ui/src/utils/CustomizePage/CustomizePagePureUtils.ts
@@ -1,0 +1,22 @@
+/*
+ *  Copyright 2026 Collate.
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+import { TAB_LABEL_MAP } from '../../constants/Customize.constants';
+import { EntityTabs } from '../../enums/entity.enum';
+import i18n from '../i18next/LocalUtil';
+
+export const getTabLabelFromId = (tab: EntityTabs): string => {
+  const labelKey = TAB_LABEL_MAP[tab];
+
+  return labelKey ? i18n.t(labelKey) : tab;
+};
