+++ openmetadata-ui/src/main/resources/ui/src/utils/Lineage/LineageUtils.tsx
@@ -10,24 +10,19 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
+export {
+  getSearchNameEsQuery,
+  prepareColumnLevelNodesFromEdges,
+  prepareDownstreamColumnLevelNodesFromDownstreamEdges,
+  prepareUpstreamColumnLevelNodesFromUpstreamEdges,
+} from './LineagePureUtils';
+
 import { ArrowLeftOutlined, ArrowRightOutlined } from '@ant-design/icons';
 import { ChevronRight } from '@untitledui/icons';
-import { get, omit, pick } from 'lodash';
 import { ReactComponent as ColumnIcon } from '../../assets/svg/ic-column.svg';
 import { ReactComponent as TableIcon } from '../../assets/svg/ic-table-new.svg';
 import { CondensedBreadcrumb } from '../../components/CondensedBreadcrumb/CondensedBreadcrumb.component';
-import {
-  ColumnLevelLineageNode,
-  EdgeDetails,
-  LineageNodeType,
-  NodeData,
-} from '../../components/Lineage/Lineage.interface';
 import { EImpactLevel } from '../../components/LineageTable/LineageTable.interface';
-import { LineageDirection } from '../../generated/api/lineage/lineageDirection';
-import { Column } from '../../generated/entity/data/table';
-import { TagLabel } from '../../generated/type/tagLabel';
-import { TableSearchSource } from '../../interface/search.interface';
-import { QueryFieldInterface } from '../../pages/ExplorePage/ExplorePage.interface';
 import i18n from '../i18next/LocalUtil';
 
 export const LINEAGE_IMPACT_OPTIONS = [
@@ -55,141 +50,6 @@ export const LINEAGE_DEPENDENCY_OPTIONS = [
     icon: <ArrowLeftOutlined />,
   },
 ];
-
-const buildColumnTagMap = <T,>(
-  entityData: T & { columns?: Column[] }
-): Map<string, TagLabel[]> => {
-  const map = new Map<string, TagLabel[]>();
-  const columns = entityData.columns;
-  if (columns) {
-    for (const col of columns) {
-      // recursive
-      if (col.children) {
-        buildColumnTagMap({
-          columns: col.children,
-        }).forEach((tags, fqn) => {
-          map.set(fqn, tags);
-        });
-      }
-      if (col.fullyQualifiedName && col.tags) {
-        map.set(col.fullyQualifiedName, col.tags);
-      }
-    }
-  }
-
-  return map;
-};
-
-export const prepareColumnLevelNodesFromEdges = (
-  edges: EdgeDetails[],
-  nodes: Record<string, LineageNodeType>,
-  direction: LineageDirection = LineageDirection.Downstream
-) => {
-  const entityKey =
-    direction === LineageDirection.Upstream ? 'fromEntity' : 'toEntity';
-
-  return edges.reduce((acc: ColumnLevelLineageNode[], node: EdgeDetails) => {
-    if ((node.columns?.length ?? 0) > 0) {
-      const entityData = get(
-        nodes[node[entityKey].fullyQualifiedName ?? ''],
-        'entity'
-      );
-      const nodeDepth = get(
-        nodes[node[entityKey].fullyQualifiedName ?? ''],
-        'nodeDepth',
-        0
-      );
-
-      if (!entityData) {
-        return acc;
-      }
-
-      const picked = pick<NodeData['entity']>(
-        entityData,
-        'owners',
-        'tier',
-        'tags',
-        'domains',
-        'description'
-      ) as Pick<
-        TableSearchSource,
-        'tags' | 'tier' | 'domains' | 'description' | 'owners' | 'id'
-      >;
-
-      // Build column FQN → tags lookup map once per entity (O(C)), O(1) per lookup
-      const columnTagMap = buildColumnTagMap(entityData);
-
-      for (const col of node.columns ?? []) {
-        // flatten the fromColumns to create separate nodes for each
-        for (const fromCol of col.fromColumns || []) {
-          // Use column-specific tags instead of table tags
-          const columnFqn =
-            direction === LineageDirection.Downstream ? col.toColumn : fromCol;
-          const columnTags = columnFqn ? columnTagMap.get(columnFqn) ?? [] : [];
-
-          acc.push({
-            ...omit(node, 'columns'),
-            fromColumn: fromCol,
-            toColumn: col.toColumn,
-            docId: fromCol + '->' + col.toColumn,
-            nodeDepth,
-            ...picked,
-            tags: columnTags,
-          } as ColumnLevelLineageNode);
-        }
-      }
-    }
-
-    return acc;
-  }, []);
-};
-
-export const prepareDownstreamColumnLevelNodesFromDownstreamEdges = (
-  edges: EdgeDetails[],
-  nodes: Record<string, LineageNodeType>
-) => {
-  return prepareColumnLevelNodesFromEdges(
-    edges,
-    nodes,
-    LineageDirection.Downstream
-  );
-};
-
-export const prepareUpstreamColumnLevelNodesFromUpstreamEdges = (
-  edges: EdgeDetails[],
-  nodes: Record<string, LineageNodeType>
-) => {
-  return prepareColumnLevelNodesFromEdges(
-    edges,
-    nodes,
-    LineageDirection.Upstream
-  );
-};
-
-export const getSearchNameEsQuery = (
-  searchText: string
-): QueryFieldInterface => {
-  return {
-    bool: {
-      should: [
-        {
-          wildcard: {
-            ['name.keyword']: {
-              value: `*${searchText}*`,
-            },
-          },
-        },
-        {
-          wildcard: {
-            ['displayName.keyword']: {
-              value: `*${searchText}*`,
-            },
-          },
-        },
-      ],
-    },
-  };
-};
 
 export const getTruncatedPath = (path: string, className?: string) => {
   if (!path) {
